Refined hyperparameters and improved training process for price forecast
The model now has more and larger layers.
The mixed branch of the model now has inputs with finer time resolution.
The loss function for training is now mean absolute error, which seems to
give better accuracy for longer term forecasts.
The training process now uses the best weights found at any point in each
run; this counteracts the tendency of the larger model to overfit.